2.5 update 2
Now with running MachineActivity!  THANK YOU ARCHIT!  TODO list: Fix the
MachineActivity View (it looks strange), set TextWatchers for all of the
important information in MachineActivity, Set position tags on the
Adapters, limit minimum deleteColumn, make MachineTypeListAdapter titles
editable.  If we have time after that, option menus for the
MachineTypeListAdapter and impliment the buttons in MachineActivity so
that we can add new maintenances.